made params double

+++ core/src/io/piotrjastrzebski/playground/tiledgentest/OpenNoise.java
@@ -37,33 +37,25 @@ public class OpenNoise {
 		}
 	}
 
-	public double getNoise (int x, int y) {
+	public double getNoise (double x, double y) {
 		double result = 0;
 		for (int i = 0; i < octaves.length; i++) {
-			//double frequency = Math.pow(2,i);
-			//double amplitude = Math.pow(persistence,octaves.length-i);
 			result = result + octaves[i].eval(x / frequencies[i], y / frequencies[i]) * amplitudes[i];
 		}
 		return result;
 	}
 
-	public double getNoise (int x, int y, int z) {
+	public double getNoise (double x, double y, double z) {
 		double result = 0;
 		for (int i = 0; i < octaves.length; i++) {
-//			double frequency = Math.pow(2, i);
-//			double amplitude = Math.pow(persistence, octaves.length - i);
-
 			result = result + octaves[i].eval(x / frequencies[i], y / frequencies[i], z / frequencies[i]) * amplitudes[i];
 		}
 		return result;
 	}
 
-	public double getNoise (int x, int y, int z, int w ) {
+	public double getNoise (double x, double y, double z, double w ) {
 		double result = 0;
 		for (int i = 0; i < octaves.length; i++) {
-//			double frequency = Math.pow(2, i);
-//			double amplitude = Math.pow(persistence, octaves.length - i);
-
 			result = result + octaves[i].eval(x / frequencies[i], y / frequencies[i], z / frequencies[i], w / frequencies[i]) * amplitudes[i];
 		}
 		return result;
